modify AttendeeContoller, to add Gate facade class to authorize the user i.e. event owner or the attendee itself, to delete the attendee from an event.

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -52,8 +53,12 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
-    {
+    public function destroy(Event $event, Attendee $attendee)
+    {   
+        if(Gate::denies('delete_attendee', [$event, $attendee] )){
+            abort(403, 'You are not authorized to delete an attendee');
+        }
+
         $attendee->delete();
 
         return response(status: 204);
